Merge points which are obvious duplicates
Some recording programs (i.e. GPSLogger for Android) produce multiple
track points which contain data describing the same datum: they have the
same location and timestamp. The points contain the information which
completes the data, like elevation in one, but xDOP values in the other.
Such trackpoints can be safely merged in one.
The code generalizes it and removes the trackpoints which don't add
any information, which is not present in the previous trackpoint.