Improve formatting and avoid potential overflow in memory allocation
The callocs first argument expects a size_t, while parallel has been an
int. Given that the arguments range is 1..SIZE_MAX/2, it is possible
for the parallel value (specifiable from command-line) to overflow
the argument. Which would make calloc to return NULL and crash the
program. Not dangerous, but annoying